+++ DTXMania/Stage/05.SongSelection/CActSelectPopupMenu.cs
@@ -329,7 +329,7 @@ internal class CActSelectPopupMenu : CActivity
                 }
                 #endregion
                 #region [ キー入力: 前に移動 ]
-                ctキー反復用.Up.tRepeatKey(CDTXMania.InputManager.Keyboard.bKeyPressing((int)SlimDXKey.UpArrow), new CCounter.DGキー処理(tMoveToPrevious));
+                ctキー反復用.Up.tRepeatKey(CDTXMania.InputManager.Keyboard.bKeyPressing(SlimDXKey.UpArrow), new CCounter.DGキー処理(tMoveToPrevious));
                 ctキー反復用.R.tRepeatKey(CDTXMania.Pad.bPressingGB(EPad.R), new CCounter.DGキー処理(tMoveToPrevious));
                 //Change to HT
                 if (CDTXMania.Pad.bPressed(EInstrumentPart.DRUMS, EPad.HT))
@@ -338,7 +338,7 @@ internal class CActSelectPopupMenu : CActivity
                 }
                 #endregion
                 #region [ キー入力: 次に移動 ]
-                ctキー反復用.Down.tRepeatKey(CDTXMania.InputManager.Keyboard.bKeyPressing((int)SlimDXKey.DownArrow), new CCounter.DGキー処理(tMoveToNext));
+                ctキー反復用.Down.tRepeatKey(CDTXMania.InputManager.Keyboard.bKeyPressing(SlimDXKey.DownArrow), new CCounter.DGキー処理(tMoveToNext));
                 ctキー反復用.B.tRepeatKey(CDTXMania.Pad.bPressingGB(EPad.G), new CCounter.DGキー処理(tMoveToNext));
                 //Change to LT
                 if (CDTXMania.Pad.bPressed(EInstrumentPart.DRUMS, EPad.LT))
